requirements-actions.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
playwright>=1.58.0
Pillow>=10.0.0
